Fix segfault in DX12 renderer
Check the value of ID3D12Resource::Map to prevent segfaulting.

It might not have to check the value of m_OcclusionDownloadBuffer->Map as m_OcclusionMemory is still nullptr when it fails when I was testing, but I can't find any guarantees that that is always the case. It's quite inconsistent to replicate but running a scene while shaders are compiling seem to be the best recipe for triggering Map to fail.